Rebuilt app in 3D
After a long break I basically restarted the whole app so it could be
switched over to full 3D. I threw out lots of the old framework like the
object and entity manager, and rewrote a lot of the existing stuff such
as the UI code. Although switching to 3D required rewriting most of the
app, it made the code much more concise and easy to read as well as
giving more freedom over what can be done (for example it makes it
really easy to change lighting without making new assets). It also
greatly speeds up asset production (since you dont have to render every
frame of animations and such).